Change: YOLO 모델 코드 수정
